abstract jwt related stuffs
