--- src/permissions/jwt_claim.rs
@@ -1,8 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Claims {
-    pub exp: u64,
-    pub iss: String,
-    pub username: String
-}
